Created the index.html boilerplate
I created the index.html boilerplate with all the required information on it.
It specifies the version of HTML, it sets the language of the document to english, and sets the characters to UTF-8
I also set the title of the page to Odin Recipes homepage and added the Odin Recipes header to the index.html.