Added image, caption, and description for recipe
In Lasagna.html, I added a caption to describe the image that was also
added in this commit. The image is a picture of lasagna taken off of the
internet. The description for recipe is a short description, saying that
the reader will enjoy the recipe, how many people it feeds, and how long
it takes to make the recipe as well.